Fix typo in docker-compose.yml.
docker-compose.yml was erroneously referring to the "mozilla-bteam" Docker
team, which should be "mozillabteam".  This resulted in an error when
trying to bring up the container.

Also fixed the docs to remove reference to building the container, since
"docker build" just reports "bugzilla uses an image, skipping".  Until this
section is updated with the proper build instructions, it should be
removed.